Include gated track tile actions menu
